pyproject.toml
[tool.poetry]
name = "txt_sink_api"
version = "0.1.0"
description = "API for Dockerized Django Web Application with AWS (Amazon Web Services) Integration"
authors = ["Adem Mohammed <abulhuman.dev@gmail.com>"]
readme = "README.rst"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.6"
django-filter = "^24.3"
boto3 = "^1.35.84"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-xdist = "^3.6.1"
model-bakery = "^1.20.0"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
split-settings = "^1.0.0"
drf-spectacular = "^0.28.0"
django-storages = "^1.14.4"
django-cors-headers = "^4.6.0"
pydantic = "^2.10.4"
pynacl = "^1.5.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
colorlog = "^6.9.0"

[tool.isort]
multi_line_output = 5
line_length = 120

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 120
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
